nix-gl-host: first relatively naive implementation
First take.

For now, we're only try to support the Nvidia proprietary driver. We
also cut quite some corners :)

We hardcode a list of DSOs we're looking for in the code. That's
obviously the best long-term decision, we'll have to revise this
particular approach later on.

We're looking for these listed DSOs in the GL_VENDOR_PATH provided by
the user. We'll need to patch these DSOs and we obviously don't want
to alter the host OS configuration. So we have to first copy them to
the user XDG cache directory.

Once copied, we alter their runpath to point to the user cache dir:
these DSO can depend on each other.

Finally, we point the patched libglvnd GLX implementation to the cache
dir and replace the current process with the target one.